pyproject.toml
[project]
name = "prompt-siren"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Edoardo Debenedetti", email = "edebenedetti@meta.com" }]
requires-python = ">=3.10,<3.14"
classifiers = [
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]
dependencies = [
    "agentdojo>=0.1.35",
    "aiohttp>=3.12.15",
    "anyio>=4.10.0",
    "click>=8.3.0",
    # Only needed for py3.10 as >=3.11 comes with
    # ExceptionGroup in the standard library
    "exceptiongroup>=1.2.0; python_version < '3.11'",
    "backports.strenum (>=1.3.1,<2.0) ; python_version < '3.11'",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "logfire>=4.7.0",
    "pandas>=2.0.0",
    "playwright>=1.54.0",
    "pydantic-ai-slim[anthropic,bedrock,openai]>=1.1.0",
    "pyyaml>=6.0.2",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "omegaconf>=2.3.0",
    "filelock>=3.19.1",
    "aiodocker>=0.24.0",
    "tabulate>=0.9.0",
    "swebench",
    "jinja2>=3.1.6",
    "typing-extensions>=4.15.0",
]

[dependency-groups]
dev = [
    "anyio>=4.10.0",
    "arize-phoenix>=11.30.0",
    "basedpyright>=1.31.7",
    "pytest>=8.4.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.9",
]

[project.scripts]
prompt-siren = "prompt_siren.cli:main"

[project.entry-points."prompt_siren.agents"]
plain = "prompt_siren.agents.plain:create_plain_agent"

[project.entry-points."prompt_siren.attacks"]
agentdojo = "prompt_siren.attacks.agentdojo_attack:create_agentdojo_attack"
dict = "prompt_siren.attacks.dict_attack:create_dict_attack"
file = "prompt_siren.attacks.dict_attack:create_dict_attack_from_file"
mini-goat = "prompt_siren.attacks.mini_goat_attack:create_mini_goat_attack"

[project.entry-points."prompt_siren.datasets"]
agentdojo = "prompt_siren.datasets.agentdojo_dataset:create_agentdojo_dataset"
swebench = "prompt_siren.datasets.swebench_dataset:create_swebench_dataset"

[project.entry-points."prompt_siren.sandbox_managers"]
local-docker = "prompt_siren.sandbox_managers.docker:create_docker_sandbox_manager"

[project.entry-points."prompt_siren.providers"]
bedrock = "prompt_siren.providers.bedrock:BedrockProvider"
llama = "prompt_siren.providers.llama:LlamaProvider"

[project.optional-dependencies]
ufmt = [
    "black==25.11.0",
    "ruff-api==0.1.0",
    "stdlibs==2024.1.28",
    "ufmt==2.8.0",
    "usort==1.0.8.post1",
]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "B",
    "A",
    "C4",
    "ICN",
    "PIE",
    "PT",
    "Q",
    "RET",
    "N",
    "PERF",
    "I",
    "UP",
    "RUF",
]

# Settings to ensure compatibility with Meta code style guides
[tool.ruff.lint.isort]
case-sensitive = false
combine-as-imports = true
detect-same-package = false
order-by-type = false
# Required as Meta's code style puts together external and
# internal imports
known-third-party = ["prompt_siren", "tests"]

# Needed for ufmt
[tool.black]
line-length = 100

[tool.ufmt]
formatter = "ruff-api"
sorter = "ruff-api"

[tool.pyright]
typeCheckingMode = "standard"

[tool.pytest.ini_options]
markers = [
    "docker_integration: marks tests as integration tests that require Docker (deselect with '-m \"not docker_integration\"')",
]
filterwarnings = [
    # Ignore TestSpec class from swebench module
    "ignore:cannot collect test class 'TestSpec' because it has a __init__ constructor:pytest.PytestCollectionWarning",
    # Warning coming from external Google module
    "ignore:Using `@model_validator` with mode='after' on a classmethod is deprecated:pydantic.PydanticDeprecatedSince212:google.genai.types",
]

[tool.uv.sources]
swebench = { git = "https://github.com/princeton-nlp/SWE-bench.git", rev = "v4.1.0" }
